refactor: Update import path for repair functions in hooks-for-nerps.js
